Commit the contents of this project anew
This project previously consisted of some 277 commits starting on May
25, 2011.  I decided to squash all commits, clean up the codebase, and
bring in both the serial and RF variants of this project.

This codebase previously did not use pull requests or thoughtful
branching, and the only contributor was me.  Therefore, I felt that the
benefits of declaring surrender to the overly-complex history and lack
of careful management when adding RF support outweighed the downsides.