Added keepalive plugin.
I have a sneaking suspicion that my serial controller might be doing
something weird because when I stop sending data and then try again a
few hours later, I have to reset the lights.  Part of the problem is
probably the lack of a pulldown resistor on the data enable line of the
SN75176B.  But this might help mitigate the issue as well by sending an
"off" packet every second.